Use long type when time spans are handled in seconds

+++ AnkiDroid/src/main/java/com/ichi2/libanki/Stats.java
@@ -764,8 +764,8 @@ public class Stats {
         }
         mFirstElement = 0;
         mMaxElements = list.size() - 1;
-        mAverage = Utils.timeSpan(context, (int) Math.round(avg * SECONDS_PER_DAY));
-        mLongest = Utils.timeSpan(context, (int) Math.round(max_ * SECONDS_PER_DAY));
+        mAverage = Utils.timeSpan(context, (long) Math.round(avg * SECONDS_PER_DAY));
+        mLongest = Utils.timeSpan(context, (long) Math.round(max_ * SECONDS_PER_DAY));
 
         //some adjustments to not crash the chartbuilding with emtpy data
         if (mMaxElements == 0) {
